[refactor] 패키지 수정에 따른 import문 변경
