+++ src/main/java/com/friedNote/friedNote_backend/domain/alarm/domain/service/AlarmSaveService.java
@@ -1,6 +1,6 @@
 package com.friedNote.friedNote_backend.domain.alarm.domain.service;
 
-import com.friedNote.friedNote_backend.common.DomainService;
+import com.friedNote.friedNote_backend.common.annotation.DomainService;
 import com.friedNote.friedNote_backend.domain.alarm.domain.entity.Alarm;
 import com.friedNote.friedNote_backend.domain.alarm.domain.repository.AlarmRepository;
 import lombok.RequiredArgsConstructor;
