+++ src/main/java/com/friedNote/friedNote_backend/domain/bookmark/application/service/BookmarkGetUseCase.java
@@ -1,6 +1,6 @@
 package com.friedNote.friedNote_backend.domain.bookmark.application.service;
 
-import com.friedNote.friedNote_backend.common.UseCase;
+import com.friedNote.friedNote_backend.common.annotation.UseCase;
 import com.friedNote.friedNote_backend.domain.bookmark.application.dto.response.BookmarkResponse;
 import com.friedNote.friedNote_backend.domain.bookmark.application.mapper.BookmarkMapper;
 import com.friedNote.friedNote_backend.domain.bookmark.domain.service.BookmarkQueryService;
