+++ src/main/java/com/friedNote/friedNote_backend/domain/bookmark/domain/entity/Bookmark.java
@@ -1,6 +1,6 @@
 package com.friedNote.friedNote_backend.domain.bookmark.domain.entity;
 
-import com.friedNote.friedNote_backend.common.BaseTimeEntity;
+import com.friedNote.friedNote_backend.common.domain.BaseTimeEntity;
 import com.friedNote.friedNote_backend.domain.recipe.domain.entity.Recipe;
 import com.friedNote.friedNote_backend.domain.user.domain.entity.User;
 import jakarta.persistence.*;
