+++ src/main/java/com/friedNote/friedNote_backend/domain/bookmark/domain/service/BookmarkQueryService.java
@@ -1,6 +1,6 @@
 package com.friedNote.friedNote_backend.domain.bookmark.domain.service;
 
-import com.friedNote.friedNote_backend.common.DomainService;
+import com.friedNote.friedNote_backend.common.annotation.DomainService;
 import com.friedNote.friedNote_backend.domain.bookmark.domain.entity.Bookmark;
 import com.friedNote.friedNote_backend.domain.bookmark.domain.repository.BookmarkRepository;
 import lombok.RequiredArgsConstructor;
