+++ src/main/java/com/friedNote/friedNote_backend/domain/cookingProcess/domain/service/CookingProcessSaveService.java
@@ -1,6 +1,6 @@
 package com.friedNote.friedNote_backend.domain.cookingProcess.domain.service;
 
-import com.friedNote.friedNote_backend.common.DomainService;
+import com.friedNote.friedNote_backend.common.annotation.DomainService;
 import com.friedNote.friedNote_backend.domain.cookingProcess.domain.entity.CookingProcess;
 import com.friedNote.friedNote_backend.domain.cookingProcess.domain.repository.CookingProcessRepository;
 import lombok.RequiredArgsConstructor;
