+++ src/main/java/com/friedNote/friedNote_backend/domain/ingredientGroup/domain/service/IngredientGroupSaveService.java
@@ -1,6 +1,6 @@
 package com.friedNote.friedNote_backend.domain.ingredientGroup.domain.service;
 
-import com.friedNote.friedNote_backend.common.DomainService;
+import com.friedNote.friedNote_backend.common.annotation.DomainService;
 import com.friedNote.friedNote_backend.domain.ingredientGroup.domain.entity.IngredientGroup;
 import com.friedNote.friedNote_backend.domain.ingredientGroup.domain.repository.IngredientGroupRepository;
 import lombok.RequiredArgsConstructor;
