+++ src/main/java/com/friedNote/friedNote_backend/domain/inventory/domain/service/InventorySaveService.java
@@ -1,6 +1,6 @@
 package com.friedNote.friedNote_backend.domain.inventory.domain.service;
 
-import com.friedNote.friedNote_backend.common.DomainService;
+import com.friedNote.friedNote_backend.common.annotation.DomainService;
 import com.friedNote.friedNote_backend.domain.inventory.domain.entity.Inventory;
 import com.friedNote.friedNote_backend.domain.inventory.domain.repository.InventoryRepository;
 import lombok.RequiredArgsConstructor;
