+++ src/main/java/com/friedNote/friedNote_backend/domain/recipe/domain/entity/Recipe.java
@@ -1,6 +1,6 @@
 package com.friedNote.friedNote_backend.domain.recipe.domain.entity;
 
-import com.friedNote.friedNote_backend.common.BaseTimeEntity;
+import com.friedNote.friedNote_backend.common.domain.BaseTimeEntity;
 import com.friedNote.friedNote_backend.domain.recipeBook.domain.entity.RecipeBook;
 import com.friedNote.friedNote_backend.domain.user.domain.entity.User;
 import jakarta.persistence.*;
