+++ src/main/java/com/friedNote/friedNote_backend/domain/recipe/domain/service/RecipeQueryService.java
@@ -1,6 +1,6 @@
 package com.friedNote.friedNote_backend.domain.recipe.domain.service;
 
-import com.friedNote.friedNote_backend.common.DomainService;
+import com.friedNote.friedNote_backend.common.annotation.DomainService;
 import com.friedNote.friedNote_backend.domain.recipe.domain.entity.Recipe;
 import com.friedNote.friedNote_backend.domain.recipe.domain.repository.RecipeRepository;
 import lombok.RequiredArgsConstructor;
