+++ src/main/java/com/friedNote/friedNote_backend/domain/recipeBook/domain/service/RecipeBookQueryService.java
@@ -1,6 +1,6 @@
 package com.friedNote.friedNote_backend.domain.recipeBook.domain.service;
 
-import com.friedNote.friedNote_backend.common.DomainService;
+import com.friedNote.friedNote_backend.common.annotation.DomainService;
 import com.friedNote.friedNote_backend.domain.recipeBook.domain.entity.RecipeBook;
 import com.friedNote.friedNote_backend.domain.recipeBook.domain.repository.RecipeBookRepository;
 import lombok.RequiredArgsConstructor;
